% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CulmenPenguins.R
\docType{data}
\name{CulmenPenguins}
\alias{CulmenPenguins}
\title{Culmen of Penguins}
\format{
A data frame with 342 rows and 3 variables:
 \describe{
     \item{Culmen.Length..mm.}{The Culmen length of Penguins in mm.}
     \item{Culmen.Depth..mm.}{The Culmen depth of Penguins in mm.}
     \item{Species}{"Adelie","Gentoo","Chinstrap"}
     }
}
\source{
{Kaggle:https://www.kaggle.com/code/tirendazacademy/penguin-dataset-data-visualization-with-seaborn}
}
\usage{
CulmenPenguins
}
\description{
The dataset Penguins with Culmen.Length(mm),Culmen.Depth(mm),and species.
The NAs of raw data has been removed.
}
\examples{
  data("CulmenPenguins")
}
\keyword{datasets}
