% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataA.R
\docType{data}
\name{DataA}
\alias{DataA}
\title{Data A}
\format{
A data frame with 490 rows and 2 variables:
 \describe{
     \item{lstat}{lstat}
     \item{medv.log}{log medv.}
     }
}
\source{
{National Tsing Hua University Institute of Statistics}
}
\usage{
DataA
}
\description{
DataA from statistical computing HW4
}
\examples{
  data("DataA")
}
\keyword{datasets}
