% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataB.R
\docType{data}
\name{DataB}
\alias{DataB}
\title{Data B}
\format{
A data frame with 100 rows and 2 variables:
 \describe{
     \item{X}{index}
     \item{x}{x}
     }
}
\source{
{National Tsing Hua University Institute of Statistics}
}
\usage{
DataB
}
\description{
DataB from statistical computing HW4
}
\examples{
  data("DataB")
}
\keyword{datasets}
