% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataC.R
\docType{data}
\name{DataC}
\alias{DataC}
\title{Data C}
\format{
A data frame with 300 rows and 3 variables:
 \describe{
     \item{X1}{X1}
     \item{X2}{X2}
     \item{X3}{X3}
     }
}
\source{
{National Tsing Hua University Institute of Statistics}
}
\usage{
DataC
}
\description{
DataC from statistical computing HW5
}
\examples{
  data("DataC")
}
\keyword{datasets}
