% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_GMM.R
\name{em_GMM}
\alias{em_GMM}
\title{GMM by EM algorithm}
\usage{
em_GMM(X, k, max.iter = 100)
}
\arguments{
\item{X}{X:n x p data matrix or data.frame with numeric elements}

\item{k}{k:the cluster you specified(nonzero integer)}

\item{max.iter}{the maximum iteration number(default is 100)}
}
\value{
A list of the result:
\item{data.matrix}{The input X}
\item{weight}{The vector of weights(k elements)}
\item{mu}{The gaussian mean by column for each cluster}
\item{sigma}{The variance-covariance matrix for each cluster}
\item{density}{The mixture density for all observations}
\item{BIC}{The value of BIC}
\item{AIC}{The value of AIC}
\item{count}{The number of iteration}
\item{logL.iter}{The log likelihood obtained by each iteration}
\item{data.dim}{The dimension of X,n x p}
\item{logL}{The last log likelihood}
\item{cluster}{The vector of n clusters assigned to each observation}
\item{success}{0:converge;1:No convergence}
}
\description{
The estimation of parameters for Gaussian mixture model by EM algorithm.
}
\details{
After the run, if the number of variants belonging to a group is too low,
 the message "The s.d.s are unstable." is displayed and the calculation is stopped,
 and the current result is recorded.
  In addition, if the specified argument "max.iter" fails to complete the convergence,
  the message "The iteration reaches maximum iterations" is displayed.
}
\examples{
  data(faithful)
  em_GMM(X=faithful,k=2,max.iter=10000)
}
