% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GMM.R
\name{plot_GMM}
\alias{plot_GMM}
\title{The plot of Gaussian Mixture model}
\usage{
plot_GMM(gm, type = "pairs")
}
\arguments{
\item{gm}{The object from em_GMM}

\item{type}{The type of plot,two choice:"hist","pairs".}
}
\value{
A plot specified type
}
\description{
Show the result of em_GMM.
}
\details{
For non-universal dataset,Only histogram with density line is displayed,
regardless of what argument type is entered.
}
